+++ logprep/util/helper.py
@@ -30,7 +30,7 @@ def print_fcolor(fore: AnsiFore, message: str):
     color_print_line(None, fore, message)
 
 
-def add_field_to(event, output_field, content, extends_lists=False):
+def add_field_to(event, output_field, content, extends_lists=False, overwrite_output_field=False):
     """
     Add content to an output_field in the given event. Output_field can be a dotted subfield.
     In case of missing fields all intermediate fields will be created.
@@ -41,18 +41,23 @@ def add_field_to(event, output_field, content, extends_lists=False):
         Original log-event that logprep is currently processing
     output_field: str
         Dotted subfield string indicating the target of the output value, e.g. destination.ip
-    content: str, dict
-        Value that should be written into the output_field, can be a str or dict object
+    content: str, list, dict
+        Value that should be written into the output_field, can be a str, list or dict object
     extends_lists: bool
-        Flag that determines whether or not lists as existing field values should be extended
+        Flag that determines whether output_field lists should be extended
+    overwrite_output_field: bool
+        Flag that determines whether the output_field should be overwritten
 
     Returns
     ------
     This method returns true if no conflicting fields were found during the process of the creation
     of the dotted subfields. If conflicting fields were found False is returned.
-
-    # code is originally from the generic adder, such that duplicated code could be removed there.
     """
+
+    assert not (
+        extends_lists & overwrite_output_field
+    ), "An output field can't be overwritten and extended at the same time"
+
     conflicting_fields = []
 
     keys = output_field.split(".")
@@ -66,16 +71,24 @@ def add_field_to(event, output_field, content, extends_lists=False):
 
         if isinstance(dict_[key], dict) and idx < len(keys) - 1:
             dict_ = dict_[key]
-        elif isinstance(dict_[key], list) and extends_lists and idx == len(keys) - 1:
+        elif (
+            isinstance(dict_[key], list)
+            and extends_lists
+            and idx == len(keys) - 1
+            and not overwrite_output_field
+        ):
+            if isinstance(content, str):
+                content = [content]
             dict_[key].extend(content)
         else:
-            conflicting_fields.append(keys[idx])
-            break
+            if not overwrite_output_field:
+                conflicting_fields.append(keys[idx])
+                break
+            dict_[key] = content
 
     if conflicting_fields:
         return False
-    else:
-        return True
+    return True
 
 
 def get_dotted_field_value(event: dict, dotted_field: str) -> Optional[Union[dict, list, str]]:
@@ -94,18 +107,59 @@ def get_dotted_field_value(event: dict, dotted_field: str) -> Optional[Union[dic
     -------
     dict_: dict, list, str
         The value of the requested dotted field.
-
-    # code is originally from the Processor, such that duplicated code could be removed there.
     """
 
     fields = dotted_field.split(".")
-    dict_ = event
-    for field in fields:
-        if field in dict_ and isinstance(dict_, dict):
-            dict_ = dict_[field]
-        else:
-            return None
-    return dict_
+    return _retrieve_field_value_and_delete_field_if_configured(
+        event, fields, delete_source_field=False
+    )
+
+
+def pop_dotted_field_value(event: dict, dotted_field: str) -> Optional[Union[dict, list, str]]:
+    """
+    Remove and return dotted field. Returns None is field does not exist.
+
+    Parameters
+    ----------
+    event: dict
+        The event from which the dotted field value should be extracted
+    dotted_field: str
+        The dotted field name which identifies the requested value
+
+    Returns
+    -------
+    dict_: dict, list, str
+        The value of the requested dotted field.
+    """
+    fields = dotted_field.split(".")
+    return _retrieve_field_value_and_delete_field_if_configured(
+        event, fields, delete_source_field=True
+    )
+
+
+def _retrieve_field_value_and_delete_field_if_configured(
+    sub_dict, dotted_fields_path, delete_source_field=False
+):
+    """
+    Iterates recursively over the given dictionary retrieving the dotted field. If set the source
+    field will be removed. When again going back up the stack trace it deletes the empty left over
+    dicts.
+    """
+    next_key = dotted_fields_path.pop(0)
+    if next_key in sub_dict and isinstance(sub_dict, dict):
+        if not dotted_fields_path:
+            field_value = sub_dict[next_key]
+            if delete_source_field:
+                del sub_dict[next_key]
+            return field_value
+        field_value = _retrieve_field_value_and_delete_field_if_configured(
+            sub_dict[next_key], dotted_fields_path, delete_source_field
+        )
+        # If remaining subdict is empty delete it
+        if not sub_dict[next_key]:
+            del sub_dict[next_key]
+        return field_value
+    return None
 
 
 def recursive_compare(test_output, expected_output):
@@ -115,7 +169,7 @@ def recursive_compare(test_output, expected_output):
     if not isinstance(test_output, type(expected_output)):
         return test_output, expected_output
 
-    elif isinstance(test_output, dict) and isinstance(expected_output, dict):
+    if isinstance(test_output, dict) and isinstance(expected_output, dict):
         if sorted(test_output.keys()) != sorted(expected_output.keys()):
             return sorted(test_output.keys()), sorted(expected_output.keys())